apps/backend/requirements.txt
# Web框架
Flask==3.0.3
Flask-CORS==4.0.1
Flask-JWT-Extended==4.6.0

# 数据库
Flask-SQLAlchemy==3.1.1
Flask-Migrate==4.0.7

# 密码加密
Werkzeug==3.0.3
bcrypt==4.1.3

# HTTP请求
requests==2.32.3
aiohttp==3.9.5

# 数据验证
marshmallow==3.21.3

# 环境变量
python-dotenv==1.0.1

# 生产服务器
gunicorn==22.0.0

# 开发工具
pytest==8.2.2
pytest-cov==5.0.0
flake8==7.1.0
black==24.4.2
mypy==1.10.1

# 加密库
cryptography==41.0.7
